+++ src/prompts/personalStatement.prompt.ts
@@ -1,0 +1,37 @@
+export const PERSONAL_STATEMENT_PROMPTS = [
+  {
+    id: 'ps1',
+    prompt: 'Some students have a background, identity, interest, or talent that is so meaningful they believe their application would be incomplete without it. If this sounds like you, then please share your story.',
+    word_count: 650
+  },
+  {
+    id: 'ps2',
+    prompt: 'The lessons we take from obstacles we encounter can be fundamental to later success. Recount a time when you faced a challenge, setback, or failure. How did it affect you, and what did you learn from the experience?',
+    word_count: 650
+  },
+  {
+    id: 'ps3',
+    prompt: 'Reflect on a time when you questioned or challenged a belief or idea. What prompted your thinking? What was the outcome?',
+    word_count: 650
+  },
+  {
+    id: 'ps4',
+    prompt: 'Reflect on something that someone has done for you that has made you happy or thankful in a surprising way. How has this gratitude affected or motivated you?',
+    word_count: 650
+  },
+  {
+    id: 'ps5',
+    prompt: 'Discuss an accomplishment, event, or realization that sparked a period of personal growth and a new understanding of yourself or others.',
+    word_count: 650
+  },
+  {
+    id: 'ps6',
+    prompt: 'Describe a topic, idea, or concept you find so engaging that it makes you lose all track of time. Why does it captivate you? What or who do you turn to when you want to learn more?',
+    word_count: 650
+  },
+  {
+    id: 'ps7',
+    prompt: 'Share an essay on any topic of your choice. It can be one you\'ve already written, one that responds to a different prompt, or one of your own design.',
+    word_count: 650
+  }
+];
